Run AtomTextTest with Robolectric
